Update tracking_metrics.py (#13)
Change definition of `max_sprint_length` to be calculated as a rolling sum (to better match the tracks in the animations). Facilitates interpretation of this measurement. `num_direction_changes` (and therefore also `pivot_rate` are also changed to be defined as rolling averages, but more so to suppress noise in the trajectory.